bug fix on project repository and edit data loader, created elaborate dtos and mapper and integration

+++ src/main/java/com/infoschool/infoschool/service/ElaborateService.java
@@ -1,7 +1,11 @@
 package com.infoschool.infoschool.service;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.infoschool.infoschool.dto.request.ElaborateRequestDto;
+import com.infoschool.infoschool.dto.response.ElaborateResponseDto;
+import com.infoschool.infoschool.mapper.ElaborateMapper;
 import com.infoschool.infoschool.model.Elaborate;
 import com.infoschool.infoschool.repository.ElaborateRepository;
 
@@ -11,40 +15,69 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 public class ElaborateService {
     
-    private final ElaborateRepository elaborateRepository;
+    @Autowired
+    private ElaborateRepository elaborateRepository;
+    @Autowired
+    private ElaborateMapper elaborateMapper;
 
-    public ElaborateService(ElaborateRepository elaborateRepository) {
-        this.elaborateRepository = elaborateRepository;
-    }
-
-    public Elaborate add(Elaborate elaborate) {
+    public ElaborateResponseDto add(ElaborateRequestDto elaborate) {
         try {
             log.info("Adding new elaborate: {}", elaborate);
-            return elaborateRepository.save(elaborate);
+            Elaborate elaborateEntity = elaborateMapper.requestToEntity(elaborate);
+            elaborateRepository.save(elaborateEntity);
+            log.info("Elaborate added successfully: {}", elaborateEntity);
+            return elaborateMapper.entityToResponse(elaborateEntity);
         } catch (Exception e) {
             log.error("Error adding elaborate: {}", elaborate, e);
             throw new RuntimeException("Error adding elaborate: " + e.getMessage(), e);
         }
     }
 
-    public Elaborate edit(Elaborate elaborate) {
+    public ElaborateResponseDto edit(ElaborateRequestDto elaborate) {
         try {
             log.info("Editing elaborate: {}", elaborate);
-            if (elaborateRepository.existsById(elaborate.getId())) {
-                log.error("Elaborate with ID {} already exists", elaborate.getId());
-                throw new RuntimeException("Elaborate with this ID already exists");
+
+            Elaborate elaborateEntity = elaborateMapper.requestToEntity(elaborate);
+            elaborateEntity.setId(elaborate.getId());
+            if (elaborateEntity.getId() == null) {
+                log.error("Elaborate ID is null");
+                throw new RuntimeException("Elaborate ID cannot be null");
             }
-            return elaborateRepository.save(elaborate);
+            Elaborate existingElaborate = elaborateRepository.findById(elaborateEntity.getId()).orElse(null);
+            if (elaborateEntity.getProject() != null) {
+                existingElaborate.setProject(elaborateEntity.getProject());   
+            }
+            if (elaborateEntity.getStudent() != null) {
+                existingElaborate.setStudent(elaborateEntity.getStudent());   
+            }
+            if (elaborateEntity.getTitle() != null) {
+                existingElaborate.setTitle(elaborateEntity.getTitle());   
+            }
+            if (elaborateEntity.getComment() != null) {
+                existingElaborate.setComment(elaborateEntity.getComment());   
+            }
+            if (elaborateEntity.getEvaluation() != 0) {
+                existingElaborate.setEvaluation(elaborateEntity.getEvaluation());   
+            }
+            elaborateRepository.save(existingElaborate);
+            log.info("Elaborate edited successfully: {}", elaborateEntity);
+            return elaborateMapper.entityToResponse(existingElaborate);
         } catch (Exception e) {
             log.error("Error editing elaborate: {}", elaborate, e);
             throw new RuntimeException("Error editing elaborate: " + e.getMessage(), e);
         }
     }
 
-    public Elaborate getById(Long id) {
+    public ElaborateResponseDto getById(Long id) {
         try {
             log.info("Fetching elaborate by ID: {}", id);
-            return elaborateRepository.findById(id).orElse(null);
+            Elaborate elaborate = elaborateRepository.findById(id).orElse(null);
+            if (elaborate == null) {
+                log.error("Elaborate with ID {} not found", id);
+                throw new RuntimeException("Elaborate not found");
+            }
+            log.info("Elaborate fetched successfully: {}", elaborate);
+            return elaborateMapper.entityToResponse(elaborate);
         } catch (Exception e) {
             log.error("Error fetching elaborate by ID: {}", id, e);
             throw new RuntimeException("Error fetching elaborate by ID: " + e.getMessage(), e);
@@ -54,6 +87,10 @@ public class ElaborateService {
     public void deleteById(Long id) {
         try {
             log.info("Deleting elaborate by ID: {}", id);
+            if (!elaborateRepository.existsById(id)) {
+                log.error("Elaborate with ID {} not found", id);
+                throw new RuntimeException("Elaborate not found");
+            }
             elaborateRepository.deleteById(id);
         } catch (Exception e) {
             log.error("Error deleting elaborate by ID: {}", id, e);
